+++ src/main/java/net/inceptioncloud/dragonfly/apps/accountmanager/AccountManagerUI.kt
@@ -1,7 +1,6 @@
 package net.inceptioncloud.dragonfly.apps.accountmanager
 
 import kotlinx.coroutines.*
-import net.inceptioncloud.dragonfly.Dragonfly
 import net.inceptioncloud.dragonfly.engine.internal.SizedImage
 import net.inceptioncloud.dragonfly.engine.switch
 import net.inceptioncloud.dragonfly.engine.widgets.assembled.BackNavigation
@@ -10,7 +9,6 @@ import net.inceptioncloud.dragonfly.ui.loader.OneTimeUILoader
 import net.inceptioncloud.dragonfly.ui.screens.MainMenuUI
 import net.minecraft.client.gui.GuiScreen
 import net.minecraft.util.ResourceLocation
-import java.lang.Double.min
 
 class AccountManagerUI(val previousScreen: GuiScreen) : GuiScreen() {
 
@@ -24,31 +22,31 @@ class AccountManagerUI(val previousScreen: GuiScreen) : GuiScreen() {
         val accounts = AccountManagerApp.accounts
 
         +Image {
-            val aspectRatio = 384.0 / 76.0
+            val aspectRatio = 384.0f / 76.0f
 
             resourceLocation = ResourceLocation("dragonflyres/logos/account-manager.png")
-            height = 70.0
+            height = 70.0f
             width = height * aspectRatio
-            x = this@AccountManagerUI.width / 2.0 - width / 2.0
-            y = 70.0
+            x = this@AccountManagerUI.width / 2 - width / 2
+            y = 70.0f
         } id "logo-account-manager"
 
         +BackNavigation {
-            x = 30.0
-            y = this@AccountManagerUI.height - height - 30.0
+            x = 30.0f
+            y = this@AccountManagerUI.height - height - 30.0f
             gui(previousScreen)
         } id "back-navigation"
 
-        val space = 50.0
-        val totalWidth = accounts.sumByDouble { AccountCard.getCardWidth(it.displayName) } +
+        val space = 50.0f
+        val totalWidth = accounts.sumByDouble { AccountCard.getCardWidth(it.displayName).toDouble() }.toFloat() +
                 AccountCard.getCardWidth("") + // minimum width for AddAccountCard
                 space * accounts.size
-        var currentX = width / 2.0 - totalWidth / 2.0
+        var currentX = width / 2 - totalWidth / 2
 
         accounts.forEachIndexed { index, it ->
             val accountCard = +AccountCard(it) {
                 x = currentX
-                y = 300.0
+                y = 300.0f
                 isSelected = AccountManagerApp.selectedAccount == it
             } id "account-$index"
 
@@ -61,7 +59,7 @@ class AccountManagerUI(val previousScreen: GuiScreen) : GuiScreen() {
 
         +AddAccountCard {
             x = currentX
-            y = 300.0
+            y = 300.0f
         } id "add-account"
     }
