+++ src/main/java/net/inceptioncloud/dragonfly/apps/cosmetics/NoCosmeticsUI.kt
@@ -2,10 +2,8 @@ package net.inceptioncloud.dragonfly.apps.cosmetics
 
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
-import net.inceptioncloud.dragonfly.Dragonfly
 import net.inceptioncloud.dragonfly.cosmetics.logic.CosmeticsManager
 import net.inceptioncloud.dragonfly.design.color.DragonflyPalette
-import net.inceptioncloud.dragonfly.engine.font.FontWeight
 import net.inceptioncloud.dragonfly.engine.internal.Alignment
 import net.inceptioncloud.dragonfly.engine.switch
 import net.inceptioncloud.dragonfly.engine.tooltip.Tooltip
@@ -27,40 +25,40 @@ class NoCosmeticsUI(
 
     override fun initGui() {
         +Rectangle {
-            x = 0.0
-            y = 0.0
-            width = this@NoCosmeticsUI.width.toDouble()
-            height = this@NoCosmeticsUI.height.toDouble()
+            x = 0.0f
+            y = 0.0f
+            width = this@NoCosmeticsUI.width.toFloat()
+            height = this@NoCosmeticsUI.height.toFloat()
             color = DragonflyPalette.foreground.brighter(0.7)
         } id "background"
 
         +Image {
-            val resolution = 1368 / 912.0
+            val resolution = 1368 / 912.0f
 
-            height = 400.0
+            height = 400.0f
             width = height * resolution
             resourceLocation = ResourceLocation("dragonflyres/vectors/waiting.png")
-            x = this@NoCosmeticsUI.width / 2.0 - width / 2.0
-            y = 50.0
+            x = this@NoCosmeticsUI.width / 2 - width / 2
+            y = 50.0f
         } id "vector"
 
         +TextField {
-            positionBelow("vector", 10.0)
+            positionBelow("vector", 10.0f)
 
             staticText = "Nothing here..."
             fontRenderer = font(Typography.HEADING_1)
             textAlignHorizontal = Alignment.CENTER
             color = DragonflyPalette.background
 
-            width = (this@NoCosmeticsUI.width / 3.0).coerceIn(400.0, 800.0)
-            x = this@NoCosmeticsUI.width / 2.0 - width / 2.0
+            width = (this@NoCosmeticsUI.width / 3.0f).coerceIn(400.0f, 800.0f)
+            x = this@NoCosmeticsUI.width / 2 - width / 2
 
             adaptHeight = true
             adaptHeight()
         } id "title"
 
         +TextField {
-            positionBelow("title", 20.0)
+            positionBelow("title", 20.0f)
 
             staticText = "It looks like you don't own any cosmetic items!\n" +
                     "Head over to the ${DragonflyPalette.accentNormal.chatCode}[Dragonfly Store]" +
@@ -70,19 +68,19 @@ class NoCosmeticsUI(
             textAlignHorizontal = Alignment.CENTER
             color = DragonflyPalette.background
 
-            width = (this@NoCosmeticsUI.width / 2.5).coerceIn(400.0, 900.0)
-            x = this@NoCosmeticsUI.width / 2.0 - width / 2.0
+            width = (this@NoCosmeticsUI.width / 2.5f).coerceIn(400.0f, 900.0f)
+            x = this@NoCosmeticsUI.width / 2 - width / 2
 
             adaptHeight = true
             adaptHeight()
         } id "text"
 
         +OutlineButton {
-            positionBelow("text", 90.0)
-            width = 350.0
-            height = 55.0
+            positionBelow("text", 90.0f)
+            width = 350.0f
+            height = 55.0f
             text = "Get your cosmetics now!"
-            x = this@NoCosmeticsUI.width / 2.0 - width / 2.0
+            x = this@NoCosmeticsUI.width / 2 - width / 2
             color = DragonflyPalette.accentNormal
             onClick {
                 openWebLink(URI("https://store.playdragonfly.net"))
@@ -90,11 +88,11 @@ class NoCosmeticsUI(
         } id "get-button"
 
         +OutlineButton {
-            positionBelow("get-button", 10.0)
-            width = 350.0
-            height = 40.0
+            positionBelow("get-button", 10.0f)
+            width = 350.0f
+            height = 40.0f
             text = "Refresh cosmetics"
-            x = this@NoCosmeticsUI.width / 2.0 - width / 2.0
+            x = this@NoCosmeticsUI.width / 2 - width / 2
             tooltip = Tooltip("Synchronize cosmetics with database", TooltipAlignment.BELOW)
             onClick {
                 if (text == "Refresh cosmetics") {
